requirements.txt
numpy
pandas
scipy
scikit-learn
torch
yfinance
requests
beautifulsoup4
matplotlib
seaborn
plotly
jupyter
ipykernel
pyyaml
tqdm
statsmodels
joblib
